Fix uint[] crash on bulk item upserts
Equipment, inventory and bank all build a parameter array from the
Amount column for their UNNEST upsert. Amount is uint, and Npgsql has
no array mapping for uint[] because PostgreSQL has no unsigned integer
type, so the insert threw:

  The CLR array type System.UInt32[] isn't supported by Npgsql

The single-row overloads pass Amount as a scalar and were unaffected,
which is why this stayed hidden: the bulk path is skipped entirely when
the collection is empty, and nothing had ever populated one. Giving new
characters starting equipment was the first thing to exercise it, and
the failure aborted character creation before starting attributes were
persisted — so new characters came out with factions and abilities but
no attributes and no gear.

Widen to long when building the array, which is lossless from uint and
matches the column: amount is bigint on all three tables. The SQL cast
moves from integer[] to bigint[] to agree. The buff services keep
integer[] — their Stacks column is already int.

+++ FishMMO-Database/FishMMO-DB/Npgsql/Services/Scene/Character/CharacterBankService.cs
@@ -183,7 +183,11 @@ namespace FishMMO.Database.Npgsql.Services
 				var versionArray = activeItems.Select(i => i.Version).ToArray();
 				var templateIdArray = activeItems.Select(i => i.TemplateID).ToArray();
 				var seedArray = activeItems.Select(i => i.Seed).ToArray();
-				var amountArray = activeItems.Select(i => i.Amount).ToArray();
+				/* Amount is uint, and Npgsql has no array mapping for uint[] — Postgres has no
+				 * unsigned integer type. A single-row insert passes it as a scalar and works,
+				 * so this only surfaces on the bulk UNNEST path. Widening to long is lossless
+				 * and matches the bigint column. */
+				var amountArray = activeItems.Select(i => (long)i.Amount).ToArray();
 
 				var sql = GetUpsertSql();
 
@@ -218,7 +222,7 @@ namespace FishMMO.Database.Npgsql.Services
 					{{2}}::bigint[],
 					{{3}}::integer[],
 					{{4}}::integer[],
-					{{5}}::integer[]
+					{{5}}::bigint[]
 				) AS u(character_id, slot, version, template_id, seed, amount)
 				ON CONFLICT (character_id, slot)
 				DO UPDATE SET
